+++ app/src/main/java/io/github/miuzarte/scrcpyforandroid/pages/ScrcpyAllOptionsScreen.kt
@@ -2013,7 +2013,17 @@ internal fun ScrcpyAllOptionsPage(
                         )
                     },
                 )
-                }
+                SwitchPreference(
+                    title = stringResource(R.string.scrcpyopt_ignore_video_encoder_constraints),
+                    summary = "--ignore-video-encoder-constraints",
+                    checked = soBundle.ignoreVideoEncoderConstraints,
+                    onCheckedChange = {
+                        soBundle = soBundle.copy(
+                            ignoreVideoEncoderConstraints = it,
+                        )
+                    },
+                )
+            }
         }
 
         if (soBundle.videoSource == "display") item {